+++ tests/testthat/test-utility-functions.R
@@ -56,3 +56,22 @@ test_that("get_lambda_idx works as expected", {
 test_that("check_for_pkg functions as expected",{
   expect_error(check_for_pkg(pkg = "nonsense-package-that-doesn't-exist"))
 })
+
+test_that("reshape_peaktable works as expected",{
+  data(pk_tab)
+  pktab_long<-reshape_peaktable(pk_tab)
+  expect_equal(ncol(pktab_long),4)
+  expect_equal(nrow(pktab_long), nrow(pk_tab)*ncol(pk_tab))
+})
+
+test_that("reshape_chroms works as expected", {
+  chrom_list_long <- reshape_chroms(Sa_pr,lambdas = "210")
+  expect_equal(ncol(chrom_list_long), 4)
+  expect_equal(nrow(chrom_list_long), length(Sa_pr)*nrow(Sa_pr[[1]]))
+  expect_equal(unique(chrom_list_long$lambda), 210)
+  expect_equal(unique(chrom_list_long$sample), names(Sa_pr))
+  chrom_list_long_subset <- reshape_chroms(Sa_pr,lambdas = "210",idx = c(1:2))
+  expect_equal(ncol(chrom_list_long_subset), 4)
+  expect_equal(nrow(chrom_list_long_subset), 2*nrow(Sa_pr[[1]]))
+  expect_equal(unique(chrom_list_long_subset$sample), names(Sa_pr)[1:2])
+})