+++ NEWS.md
@@ -1,6 +1,8 @@
 # chromatographR 0.4.8 dev
 
 * Fixed `merge_peaks` function so it works properly (to combine 2 or more peaks in peak table).
+* Fixed bugs in `plot_chroms` preventing plotting with `ggplot2` and plotting wrong chromatograms in base R. 
+* Added additional tests for `plot_chroms`, reshape functions.
 
 # chromatographR 0.4.7